+++ src/ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell/public/Invoke-DotnetBuildWithRetry.ps1
@@ -107,7 +107,7 @@ function Invoke-DotnetBuildWithRetry {
 
   begin {
     $fn = 'Invoke-DotnetBuildWithRetry'
-    $mn = 'ATAP.Utilities.BuildTooling.PowerShell'
+    $mn = 'ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell'
     Write-PSFMessage -FunctionName $fn -ModuleName $mn -Level Debug -Message "Entering function $fn"
 
     try {