+++ src/ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell/public/Invoke-ModuleBuildWithRetry.ps1
@@ -132,7 +132,7 @@ function Invoke-ModuleBuildWithRetry {
 
   begin {
     $fn = 'Invoke-ModuleBuildWithRetry'
-    $mn = 'ATAP.Utilities.BuildTooling.PowerShell'
+    $mn = 'ATAP.Utilities.BuildTooling.DotnetBuild.PowerShell'
     Write-PSFMessage -FunctionName $fn -ModuleName $mn -Level Debug -Message "Entering function $fn"
 
     if (-not (Get-Command -Name 'Get-ParameterValueFromNeoConfigurationRoot' -CommandType Function -ErrorAction SilentlyContinue)) {